Restructure client so that AST generation is done at the client index.js
