+++ package.json
@@ -51,6 +51,7 @@
     "morgan": "^1.9.0",
     "normalizr": "^3.2.4",
     "pg": "^7.4.0",
+    "pluralize": "^7.0.0",
     "react": "16.0.0",
     "react-apollo": "^2.0.1",
     "react-native": "^0.50.0",
