--- src/server/graphql/client/gqlNormalizer/sampleMutationResult.json
@@ -1,145 +1,0 @@
-{
-  "data": {
-    "updateUser": {
-      "id": "1",
-      "username": "the hugest",
-      "sessions": [
-        {
-          "id": "1",
-          "name": "Study Session 1",
-          "start": "2017-10-21T15:51:09.489-07:00",
-          "activities": [
-            {
-              "id": "1",
-              "name": "Write seed data"
-            },
-            {
-              "id": "2",
-              "name": "Lunch"
-            },
-            {
-              "id": "3",
-              "name": "Poop"
-            },
-            {
-              "id": "4",
-              "name": "Write test cases"
-            },
-            {
-              "id": "5",
-              "name": "Cheer up girlfriend"
-            },
-            {
-              "id": "6",
-              "name": "Reddit"
-            },
-            {
-              "id": "7",
-              "name": "Write seed data"
-            },
-            {
-              "id": "8",
-              "name": "Cheer up girlfriend"
-            }
-          ]
-        },
-        {
-          "id": "2",
-          "name": "Study Session 2",
-          "start": "2017-10-21T15:51:09.489-07:00",
-          "activities": [
-            {
-              "id": "9",
-              "name": "Write seed data"
-            },
-            {
-              "id": "10",
-              "name": "Lunch"
-            },
-            {
-              "id": "11",
-              "name": "Poop"
-            },
-            {
-              "id": "12",
-              "name": "Write test cases"
-            },
-            {
-              "id": "13",
-              "name": "Cheer up girlfriend"
-            },
-            {
-              "id": "14",
-              "name": "Reddit"
-            },
-            {
-              "id": "15",
-              "name": "Write seed data"
-            },
-            {
-              "id": "16",
-              "name": "Cheer up girlfriend"
-            }
-          ]
-        },
-        {
-          "id": "3",
-          "name": "Study Session 3",
-          "start": "2017-10-21T15:51:09.489-07:00",
-          "activities": [
-            {
-              "id": "17",
-              "name": "Write seed data"
-            },
-            {
-              "id": "18",
-              "name": "Lunch"
-            },
-            {
-              "id": "19",
-              "name": "Poop"
-            },
-            {
-              "id": "20",
-              "name": "Write test cases"
-            },
-            {
-              "id": "21",
-              "name": "Cheer up girlfriend"
-            },
-            {
-              "id": "22",
-              "name": "Reddit"
-            },
-            {
-              "id": "23",
-              "name": "Write seed data"
-            },
-            {
-              "id": "24",
-              "name": "Cheer up girlfriend"
-            }
-          ]
-        },
-        {
-          "id": "7",
-          "name": "new session",
-          "start": "2017-10-21T15:51:09.489-07:00",
-          "activities": []
-        },
-        {
-          "id": "8",
-          "name": "new session",
-          "start": "2017-10-21T15:51:09.489-07:00",
-          "activities": []
-        },
-        {
-          "id": "9",
-          "name": "new session",
-          "start": "2017-10-21T15:51:09.489-07:00",
-          "activities": []
-        }
-      ]
-    }
-  }
-}
