+++ src/server/graphql/client/index.js
@@ -1,16 +1,23 @@
 const {normalize} = require('normalizr');
 const request = require('request-promise');
-
-// TODO: Remove this once you've got the actual query portion of the client complete.
-const queryResult = require('./gqlNormalizer/sampleQueryResult.json');
-const mutationResult = require('./gqlNormalizer/sampleMutationResult.json');
-const normalizeGql = require('./gqlNormalizer');
+const normalizeGql = require('./normalizeGql');
+const {graphql} = require('graphql');
 const {parse} = require('graphql/language/parser');
+const {introspectionQuery} = require('graphql/utilities');
+const gqlSchema = require('../schema');
 
 module.exports = {
+
+  // TODO: Consider being able to dynamically determine if query or mutation.
   query: async (query, options = {}) => {
-    const normalizrSchema = await normalizeGql(query);
+    const schemaDocumentWhole = await graphql(gqlSchema, introspectionQuery);
+    const schemaDoc = schemaDocumentWhole.data.__schema;
+    const gqlOperationAST = parse(query);
+    
     // TODO: Determine if we need to query the gql server or if we have the data we need.
+    // const shouldMakeRequest;
+
+    const normalizrSchema = await normalizeGql(gqlOperationAST, schemaDoc);
 
     const requestOptions = {
       method: 'POST',
@@ -23,28 +30,6 @@ module.exports = {
 
     try {
       const gqlResponse = await request(requestOptions);
-      const responseObject = JSON.parse(gqlResponse);
-      const normalizedData = normalize(responseObject, normalizrSchema);
-
-      return normalizedData;
-    } catch(error) {
-      console.log(error);
-    }
-  },
-
-  mutate: async (mutation, options = {}) => {
-    const normalizrSchema = await normalizeGql(mutation);
-    const requestOptions = {
-      method: 'POST',
-      uri: 'http://localhost:3000/graphql',
-      body: JSON.stringify({query: mutation}),
-      headers: {
-        'Content-Type':'application/json',
-      },
-    };
-
-    try {
-      const gqlResponse = await request(requestOptions);  
       const responseObject = JSON.parse(gqlResponse);
       const normalizedData = normalize(responseObject, normalizrSchema);
